+++ ampltools/ampltools/__init__.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
-__version__ = "0.4.3"
+__version__ = "0.4.4"
 from .notebooks import (
     ampl_notebook,
 )
 
 from .modules import (
-    add_to_path,
-    install as install_modules,
     load as load_modules,
 )
 
@@ -16,4 +14,6 @@ from .licenses import (
 
 from .utils import (
     cloud_platform_name,
+    register_magics,
+    add_to_path,
 )