+++ ampltools/ampltools/modules/__init__.py
@@ -2,12 +2,13 @@
 __version__ = "0.4.0"
 
 from .amplpypi import (
-    add_to_path,
+    path,
     install_modules as install,
     uninstall_modules as uninstall,
     load_modules as load,
     installed_modules as installed,
     available_modules as available,
+    generate_requirements as requirements,
 )
 
 from .commands import _main