+++ ampltools/ampltools/modules/commands.py
@@ -4,34 +4,48 @@ from .amplpypi import (
     uninstall_modules,
     installed_modules,
     available_modules,
+    generate_requirements,
+    load_modules,
+    path,
 )
+import subprocess
 
 
 def _main(args):
     assert len(args) >= 2
-    if args[1] == "install":
-        modules = [m for m in args[2:] if not m.startswith("-")]
-        options = [o for o in args[2:] if o.startswith("-")]
+    command, args = args[1].lower(), args[2:]
+    if command == "install":
+        modules = [m for m in args if not m.startswith("-")]
+        options = [o for o in args if o.startswith("-")]
         install_modules(modules=modules, options=options, verbose=True)
-    elif args[1] == "uninstall":
-        modules = [m for m in args[2:] if not m.startswith("-")]
-        options = [o for o in args[2:] if o.startswith("-")]
+    elif command == "uninstall":
+        modules = [m for m in args if not m.startswith("-")]
+        options = [o for o in args if o.startswith("-")]
         uninstall_modules(modules=modules, options=options, verbose=True)
-    elif args[1] in ("list", "installed"):
+    elif command in ("list", "installed"):
         names = installed_modules()
         if names == []:
             raise Exception("Could not find any modules installed.")
         print("You have the following modules installed:")
         for name in sorted(set(names)):
             print(f"\t{name}")
-    elif args[1] == "available":
+    elif command == "available":
         names = available_modules()
         if names == []:
             raise Exception("Could not find any modules for download.")
         print("You can install any of the following modules:")
         for name in sorted(set(names)):
             print(f"\t{name}")
+    elif command == "path":
+        modules = [m for m in args if not m.startswith("-")]
+        print(path(modules))
+    elif command == "run":
+        load_modules()
+        subprocess.run(" ".join(args), shell=True)
+    elif command == "requirements":
+        modules = [m for m in args if not m.startswith("-")]
+        print(generate_requirements(modules))
     else:
         raise Exception(
-            "Invalid command! Valid commands: install, uninstall, list/installed, available."
+            "Invalid command! Valid commands: install, uninstall, list/installed, available, path, requirements, run."
         )