+++ ampltools/CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.4.4 - 2023-01-20
+- Remove install_modules from ampltools.
+- Move add_to_path to ampltools.
+- Add PATH and run commands for modules.
+
 ## 0.4.3 - 2023-01-19
 - Fix bug in tools.modules.uninstall.